# rebuild speedup